+++ lib/shrine/plugins/remove_invalid.rb
@@ -1,0 +1,22 @@
+class Shrine
+  module Plugins
+    # The remove_invalid plugin automatically deletes a cached file if it was
+    # invalid and deassigns it from the record.
+    #
+    #     plugin :remove_invalid
+    module RemoveInvalid
+      module AttacherMethods
+        def validate
+          super
+        ensure
+          if errors.any?
+            delete!(get, phase: :invalid)
+            _set(nil)
+          end
+        end
+      end
+    end
+
+    register_plugin(:remove_invalid, RemoveInvalid)
+  end
+end